requirements.txt
python-telegram-bot==13.0
Pillow==8.0.0